Moving parameters class Signed-off-by: Jennifer Madiedo <jmadiedo@outlook.com>
