+++ src/CarbonAware.CLI/src/CarbonAware.CLI.csproj
@@ -17,6 +17,7 @@
     <PackageReference Include="OpenTelemetry" Version="1.4.0-rc.1" />
     <PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.0.0-rc9.10" />
     <PackageReference Include="OpenTelemetry.Exporter.Console" Version="1.4.0-rc.1" />
+    <PackageReference Include="System.IO.Filesystem" Version="4.3.0" />
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
   </ItemGroup>
